Initial configuration
This initial commit includes the main project structure with some
development tools.

There's a docker-compose file inside dockerdev, ready to deploy the full
development environment with mongodb and redis services and with the
celery worker running in a container.

The celery app is defined insithe de alerts directory that's a python
module.